layout/inspector/tests/test_bug806192.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=806192
-->
<head>
  <title>Test for Bug 806192</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<pre id="test">
<script type="application/javascript">

const utils = SpecialPowers.Cc["@mozilla.org/inspector/dom-utils;1"]
.getService(SpecialPowers.Ci.inIDOMUtils);

try {
    var res = utils.getBindingURLs({});
    ok(false, "didn't get error");
}
catch(e) { ok(true, "got expected exception"); }
</script>
</pre>
</body>
</html>
